chore: Extract common file logging implementation into BaseFileLogger
