+++ GPSTest/src/main/java/com/android/gpstest/io/BaseFileLogger.java
@@ -1,0 +1,235 @@
+package com.android.gpstest.io;
+
+import android.content.Context;
+import android.os.Environment;
+import android.util.Log;
+import android.widget.Toast;
+
+import com.android.gpstest.Application;
+import com.android.gpstest.R;
+
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+/**
+ * A base implementation of a GNSS logger to store information to a file. Originally from https://github.com/google/gps-measurement-tools/tree/master/GNSSLogger,
+ * modified for GPSTest.
+ */
+public abstract class BaseFileLogger implements FileLogger {
+
+    protected final String TAG = this.getClass().getName();
+    protected static final String FILE_PREFIX = "gnss_log";
+
+    protected final Context context;
+
+    protected final Object fileLock = new Object();
+    protected BufferedWriter fileWriter;
+    protected File file;
+    protected boolean isStarted = false;
+    protected File baseDirectory;
+
+    public BaseFileLogger(Context context) {
+        this.context = context;
+    }
+
+    public File getFile() {
+        return file;
+    }
+
+    public File getBaseDirectory() {
+        return baseDirectory;
+    }
+
+    /**
+     * Defines the file extension to be used in implementations, following the ".". So "json" would
+     * be used for files with the ".json" extension.
+     * @return the file extension to be used, following the ".". So "json" would
+     *      * be used for files with the ".json" extension.
+     */
+    abstract String getFileExtension();
+
+    /**
+     * Initialize file by adding a header, if desired for the given implementation
+     *
+     * @param writer   writer to use when writing file
+     * @param filePath path to the current file
+     */
+    abstract void writeFileHeader(BufferedWriter writer, String filePath);
+
+    /**
+     * Called after files have finished initialing within startLog() but prior to returning from startLog(), if
+     * additional init is required for a specific file logging implementation
+     * @param fileWriter
+     * @return true if the operation was successful, false if it was not
+     */
+    abstract boolean postFileInit(BufferedWriter fileWriter);
+
+    /**
+     * Start a file logging process
+     *
+     * @param existingFile The existing file if file logging is to be continued, or null if a
+     *                     new file should be created.
+     * @param date The date and time to use for the file name
+     * @return true if a new file was created, false if an existing file was used
+     */
+    public boolean startLog(File existingFile, Date date) {
+        boolean isNewFile = false;
+        synchronized (fileLock) {
+            String state = Environment.getExternalStorageState();
+            if (Environment.MEDIA_MOUNTED.equals(state)) {
+                baseDirectory = new File(Environment.getExternalStorageDirectory(), FILE_PREFIX);
+                baseDirectory.mkdirs();
+            } else if (Environment.MEDIA_MOUNTED_READ_ONLY.equals(state)) {
+                logError("Cannot write to external storage.");
+                return false;
+            } else {
+                logError("Cannot read external storage.");
+                return false;
+            }
+
+            String currentFilePath;
+
+            if (existingFile != null) {
+                // Use existing file
+                currentFilePath = existingFile.getAbsolutePath();
+                BufferedWriter writer;
+                try {
+                    writer = new BufferedWriter(new FileWriter(existingFile));
+                } catch (IOException e) {
+                    logException("Could not open file: " + currentFilePath, e);
+                    return false;
+                }
+                if (fileWriter != null) {
+                    try {
+                        fileWriter.close();
+                    } catch (IOException e) {
+                        logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
+                        return false;
+                    }
+                }
+                file = existingFile;
+                fileWriter = writer;
+            } else {
+                // Create new logging file
+                SimpleDateFormat formatter = new SimpleDateFormat("yyy_MM_dd_HH_mm_ss");
+                String fileName = String.format("%s_%s." + getFileExtension(), FILE_PREFIX, formatter.format(date));
+                File currentFile = new File(baseDirectory, fileName);
+                currentFilePath = currentFile.getAbsolutePath();
+                BufferedWriter writer;
+                try {
+                    writer = new BufferedWriter(new FileWriter(currentFile));
+                } catch (IOException e) {
+                    logException("Could not open file: " + currentFilePath, e);
+                    return false;
+                }
+
+                writeFileHeader(writer, currentFilePath);
+
+                if (fileWriter != null) {
+                    try {
+                        fileWriter.close();
+                    } catch (IOException e) {
+                        logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
+                        return false;
+                    }
+                }
+
+                file = currentFile;
+                fileWriter = writer;
+
+                Log.d(TAG, Application.get().getString(R.string.logging_to_new_file, currentFilePath));
+                isNewFile = true;
+            }
+
+
+            File file;
+            if (existingFile != null) {
+                // Use existing file
+                currentFilePath = existingFile.getAbsolutePath();
+                file = existingFile;
+            } else {
+                // Create new logging file
+                SimpleDateFormat formatter = new SimpleDateFormat("yyy_MM_dd_HH_mm_ss");
+                String fileName = String.format("%s_%s." + getFileExtension(), FILE_PREFIX, formatter.format(date));
+                file = new File(baseDirectory, fileName);
+                currentFilePath = file.getAbsolutePath();
+            }
+
+            BufferedWriter writer;
+            try {
+                writer = new BufferedWriter(new FileWriter(file));
+            } catch (IOException e) {
+                logException("Could not open file: " + currentFilePath, e);
+                return false;
+            }
+
+            if (existingFile == null) {
+                // Only if file didn't previously exist
+                writeFileHeader(writer, currentFilePath);
+            }
+
+            if (fileWriter != null) {
+                try {
+                    fileWriter.close();
+                } catch (IOException e) {
+                    logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
+                    return false;
+                }
+            }
+
+            fileWriter = writer;
+            this.file = file;
+
+            boolean postInit = postFileInit(fileWriter);
+            if (!postInit) {
+                return false;
+            }
+
+            if (existingFile == null) {
+                // Only if file didn't previously exist
+                Log.d(TAG, Application.get().getString(R.string.logging_to_new_file, currentFilePath));
+                isNewFile = true;
+            }
+
+            isStarted = true;
+        }
+        return isNewFile;
+    }
+
+    /**
+     * Returns true if the logger is already started, or false if it is not
+     *
+     * @return
+     */
+    public synchronized boolean isStarted() {
+        return isStarted;
+    }
+
+    public void close() {
+        if (fileWriter != null) {
+            try {
+                fileWriter.flush();
+                fileWriter.close();
+                fileWriter = null;
+                isStarted = false;
+            } catch (IOException e) {
+                logException("Unable to close all file streams.", e);
+                return;
+            }
+        }
+    }
+
+    protected void logException(String errorMessage, Exception e) {
+        Log.e(TAG, errorMessage, e);
+        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show();
+    }
+
+    protected void logError(String errorMessage) {
+        Log.e(TAG, errorMessage);
+        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show();
+    }
+}
