+++ GPSTest/src/main/java/com/android/gpstest/io/CsvFileLogger.java
@@ -26,10 +26,7 @@ import android.location.GnssNavigationMessage;
 import android.location.Location;
 import android.location.LocationManager;
 import android.os.Build;
-import android.os.Environment;
 import android.os.SystemClock;
-import android.util.Log;
-import android.widget.Toast;
 
 import androidx.annotation.RequiresApi;
 
@@ -39,179 +36,42 @@ import com.android.gpstest.R;
 import com.android.gpstest.util.IOUtils;
 
 import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.Locale;
 
 /**
- * A GNSS logger to store information to a file. Originally from https://github.com/google/gps-measurement-tools/tree/master/GNSSLogger,
+ * A GNSS logger to store information to a CSV file. Originally from https://github.com/google/gps-measurement-tools/tree/master/GNSSLogger,
  * modified for GPSTest.
  */
-public class CsvFileLogger implements FileLogger {
+public class CsvFileLogger extends BaseFileLogger implements FileLogger {
 
-    private static final String TAG = "FileLogger";
-    private static final String FILE_PREFIX = "gnss_log";
     private static final String COMMENT_START = "# ";
     private static final char RECORD_DELIMITER = ',';
     private static final String VERSION_TAG = "GPSTest version: ";
 
-    private final Context context;
-
-    private final Object fileLock = new Object();
-    private BufferedWriter fileWriter;
-    private File file;
-    private boolean isStarted = false;
-    private File baseDirectory;
-
     public CsvFileLogger(Context context) {
-        this.context = context;
+        super(context);
     }
 
-    public File getFile() {
-        return file;
+    @Override
+    String getFileExtension() {
+        return "csv";
     }
 
-    public File getBaseDirectory() {
-        return baseDirectory;
-    }
-
-    /**
-     * Start a file logging process
-     *
-     * @param existingFile The existing file if file logging is to be continued, or null if a
-     *                     new file should be created.
-     * @param date The date and time to use for the file name
-     * @return true if a new file was created, false if an existing file was used
-     */
-    public boolean startLog(File existingFile, Date date) {
-        boolean isNewFile = false;
-        synchronized (fileLock) {
-            String state = Environment.getExternalStorageState();
-            if (Environment.MEDIA_MOUNTED.equals(state)) {
-                baseDirectory = new File(Environment.getExternalStorageDirectory(), FILE_PREFIX);
-                baseDirectory.mkdirs();
-            } else if (Environment.MEDIA_MOUNTED_READ_ONLY.equals(state)) {
-                logError("Cannot write to external storage.");
-                return false;
-            } else {
-                logError("Cannot read external storage.");
-                return false;
-            }
-
-            String currentFilePath;
-
-            if (existingFile != null) {
-                // Use existing file
-                currentFilePath = existingFile.getAbsolutePath();
-                BufferedWriter writer;
-                try {
-                    writer = new BufferedWriter(new FileWriter(existingFile));
-                } catch (IOException e) {
-                    logException("Could not open file: " + currentFilePath, e);
-                    return false;
-                }
-                if (fileWriter != null) {
-                    try {
-                        fileWriter.close();
-                    } catch (IOException e) {
-                        logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
-                        return false;
-                    }
-                }
-                file = existingFile;
-                fileWriter = writer;
-            } else {
-                // Create new logging file
-                SimpleDateFormat formatter = new SimpleDateFormat("yyy_MM_dd_HH_mm_ss");
-                String fileName = String.format("%s_%s.txt", FILE_PREFIX, formatter.format(date));
-                File currentFile = new File(baseDirectory, fileName);
-                currentFilePath = currentFile.getAbsolutePath();
-                BufferedWriter writer;
-                try {
-                    writer = new BufferedWriter(new FileWriter(currentFile));
-                } catch (IOException e) {
-                    logException("Could not open file: " + currentFilePath, e);
-                    return false;
-                }
-
-                writeFileHeader(writer, currentFilePath);
-
-                if (fileWriter != null) {
-                    try {
-                        fileWriter.close();
-                    } catch (IOException e) {
-                        logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
-                        return false;
-                    }
-                }
-
-                file = currentFile;
-                fileWriter = writer;
-
-                Log.d(TAG, Application.get().getString(R.string.logging_to_new_file, currentFilePath));
-                isNewFile = true;
-            }
-
-
-            File file;
-            if (existingFile != null) {
-                // Use existing file
-                currentFilePath = existingFile.getAbsolutePath();
-                file = existingFile;
-            } else {
-                // Create new logging file
-                SimpleDateFormat formatter = new SimpleDateFormat("yyy_MM_dd_HH_mm_ss");
-                String fileName = String.format("%s_%s.txt", FILE_PREFIX, formatter.format(date));
-                file = new File(baseDirectory, fileName);
-                currentFilePath = file.getAbsolutePath();
-            }
-
-            BufferedWriter writer;
-            try {
-                writer = new BufferedWriter(new FileWriter(file));
-            } catch (IOException e) {
-                logException("Could not open file: " + currentFilePath, e);
-                return false;
-            }
-
-            if (existingFile == null) {
-                // Only if file didn't previously exist
-                writeFileHeader(writer, currentFilePath);
-            }
-
-            if (fileWriter != null) {
-                try {
-                    fileWriter.close();
-                } catch (IOException e) {
-                    logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
-                    return false;
-                }
-            }
-
-            fileWriter = writer;
-            this.file = file;
-
-            if (existingFile == null) {
-                // Only if file didn't previously exist
-                Log.d(TAG, Application.get().getString(R.string.logging_to_new_file, currentFilePath));
-                isNewFile = true;
-            }
-
-            isStarted = true;
-        }
-        return isNewFile;
+    @Override
+    boolean postFileInit(BufferedWriter fileWriter) {
+        // No-op for CSV files
+        return true;
     }
 
     /**
-     * Initialize file by adding a header
+     * Initialize file by adding a CSV header
      *
      * @param writer   writer to use when writing file
      * @param filePath path to the current file
      */
-    private void writeFileHeader(BufferedWriter writer, String filePath) {
+    @Override
+    void writeFileHeader(BufferedWriter writer, String filePath) {
         try {
             writer.write(COMMENT_START);
             writer.newLine();
@@ -302,29 +162,6 @@ public class CsvFileLogger implements FileLogger {
         } catch (IOException e) {
             logException(Application.get().getString(R.string.could_not_initialize_file, filePath), e);
             return;
-        }
-    }
-
-    /**
-     * Returns true if the logger is already started, or false if it is not
-     *
-     * @return
-     */
-    public synchronized boolean isStarted() {
-        return isStarted;
-    }
-
-    public void close() {
-        if (fileWriter != null) {
-            try {
-                fileWriter.flush();
-                fileWriter.close();
-                fileWriter = null;
-                isStarted = false;
-            } catch (IOException e) {
-                logException("Unable to close all file streams.", e);
-                return;
-            }
         }
     }
 
@@ -470,15 +307,5 @@ public class CsvFileLogger implements FileLogger {
                         measurement.hasCarrierFrequencyHz() ? measurement.getCarrierFrequencyHz() : "");
         fileWriter.write(measurementStream);
         fileWriter.newLine();
-    }
-
-    private void logException(String errorMessage, Exception e) {
-        Log.e(TAG, errorMessage, e);
-        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show();
-    }
-
-    private void logError(String errorMessage) {
-        Log.e(TAG, errorMessage);
-        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show();
     }
 }
