+++ GPSTest/src/main/java/com/android/gpstest/io/JsonFileLogger.java
@@ -19,9 +19,6 @@ package com.android.gpstest.io;
 import android.content.Context;
 import android.location.GnssAntennaInfo;
 import android.os.Build;
-import android.os.Environment;
-import android.util.Log;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
@@ -33,41 +30,44 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
 
 /**
- * A GNSS logger to store information to a file. Originally from https://github.com/google/gps-measurement-tools/tree/master/GNSSLogger,
+ * A GNSS logger to store information to a JSON file. Originally from https://github.com/google/gps-measurement-tools/tree/master/GNSSLogger,
  * modified for GPSTest.
  */
-public class JsonFileLogger implements FileLogger {
+public class JsonFileLogger extends BaseFileLogger implements FileLogger {
 
-    private static final String TAG = "FileLogger";
-    private static final String FILE_PREFIX = "gnss_log";
-
-    private final Context context;
-
-    private final Object fileLock = new Object();
-    private BufferedWriter fileWriter;
-    private File file;
-    private boolean isStarted = false;
-    File baseDirectory;
     ObjectMapper mapper;
     JsonGenerator jsonGenerator;
 
     public JsonFileLogger(Context context) {
-        this.context = context;
+        super(context);
     }
 
-    public File getFile() {
-        return file;
+    @Override
+    String getFileExtension() {
+        return "json";
     }
 
-    public File getBaseDirectory() {
-        return baseDirectory;
+    @Override
+    void writeFileHeader(BufferedWriter writer, String filePath) {
+        // No-op for JSON files
+    }
+
+    @Override
+    boolean postFileInit(BufferedWriter fileWriter) {
+        if (jsonGenerator == null) {
+            try {
+                jsonGenerator = mapper.getFactory().createGenerator(fileWriter);
+            } catch (IOException e) {
+                logException(Application.get().getString(R.string.unable_to_open_json_generator), e);
+                return false;
+            }
+        }
+        return true;
     }
 
     /**
@@ -75,155 +75,28 @@ public class JsonFileLogger implements FileLogger {
      *
      * @param existingFile The existing file if file logging is to be continued, or null if a
      *                     new file should be created.
+     * @param date The date and time to use for the file name
      * @return true if a new file was created, false if an existing file was used
      */
+    @Override
     public boolean startLog(File existingFile, Date date) {
-        boolean isNewFile = false;
-        synchronized (fileLock) {
-            String state = Environment.getExternalStorageState();
-            if (mapper == null) {
-                mapper = new ObjectMapper();
-            }
-            if (Environment.MEDIA_MOUNTED.equals(state)) {
-                baseDirectory = new File(Environment.getExternalStorageDirectory(), FILE_PREFIX);
-                baseDirectory.mkdirs();
-            } else if (Environment.MEDIA_MOUNTED_READ_ONLY.equals(state)) {
-                logError("Cannot write to external storage.");
-                return false;
-            } else {
-                logError("Cannot read external storage.");
-                return false;
-            }
-
-            String currentFilePath;
-
-            if (existingFile != null) {
-                // Use existing file
-                currentFilePath = existingFile.getAbsolutePath();
-                BufferedWriter writer;
-                try {
-                    writer = new BufferedWriter(new FileWriter(existingFile));
-                } catch (IOException e) {
-                    logException("Could not open file: " + currentFilePath, e);
-                    return false;
-                }
-                if (fileWriter != null) {
-                    try {
-                        fileWriter.close();
-                    } catch (IOException e) {
-                        logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
-                        return false;
-                    }
-                }
-                file = existingFile;
-                fileWriter = writer;
-            } else {
-                // Create new logging file
-                SimpleDateFormat formatter = new SimpleDateFormat("yyy_MM_dd_HH_mm_ss");
-                String fileName = String.format("%s_%s.json", FILE_PREFIX, formatter.format(date));
-                File currentFile = new File(baseDirectory, fileName);
-                currentFilePath = currentFile.getAbsolutePath();
-                BufferedWriter writer;
-                try {
-                    writer = new BufferedWriter(new FileWriter(currentFile));
-                } catch (IOException e) {
-                    logException("Could not open file: " + currentFilePath, e);
-                    return false;
-                }
-
-                if (fileWriter != null) {
-                    try {
-                        fileWriter.close();
-                    } catch (IOException e) {
-                        logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
-                        return false;
-                    }
-                }
-
-                file = currentFile;
-                fileWriter = writer;
-
-                Log.d(TAG, Application.get().getString(R.string.logging_to_new_file, currentFilePath));
-                isNewFile = true;
-            }
-
-
-            File file;
-            if (existingFile != null) {
-                // Use existing file
-                currentFilePath = existingFile.getAbsolutePath();
-                file = existingFile;
-            } else {
-                // Create new logging file
-                SimpleDateFormat formatter = new SimpleDateFormat("yyy_MM_dd_HH_mm_ss");
-                String fileName = String.format("%s_%s.txt", FILE_PREFIX, formatter.format(new Date()));
-                file = new File(baseDirectory, fileName);
-                currentFilePath = file.getAbsolutePath();
-            }
-
-            BufferedWriter writer;
-            try {
-                writer = new BufferedWriter(new FileWriter(file));
-            } catch (IOException e) {
-                logException("Could not open file: " + currentFilePath, e);
-                return false;
-            }
-
-            if (fileWriter != null) {
-                try {
-                    fileWriter.close();
-                } catch (IOException e) {
-                    logException(Application.get().getString(R.string.unable_to_close_all_file_streams), e);
-                    return false;
-                }
-            }
-
-            fileWriter = writer;
-            this.file = file;
-
-            if (jsonGenerator == null) {
-                try {
-                    jsonGenerator = mapper.getFactory().createGenerator(fileWriter);
-                } catch (IOException e) {
-                    logException(Application.get().getString(R.string.unable_to_open_json_generator), e);
-                    return false;
-                }
-            }
-
-            if (existingFile == null) {
-                // Only if file didn't previously exist
-                Log.d(TAG, Application.get().getString(R.string.logging_to_new_file, currentFilePath));
-                isNewFile = true;
-            }
-
-            isStarted = true;
+        if (mapper == null) {
+            mapper = new ObjectMapper();
         }
-        return isNewFile;
+        return super.startLog(existingFile, date);
     }
 
-    /**
-     * Returns true if the logger is already started, or false if it is not
-     *
-     * @return
-     */
-    public synchronized boolean isStarted() {
-        return isStarted;
-    }
-
+    @Override
     public void close() {
         if (fileWriter != null) {
             try {
                 jsonGenerator.close();
-                fileWriter.flush();
-                fileWriter.close();
-                fileWriter = null;
                 mapper = null;
-                isStarted = false;
             } catch (IOException e) {
-                logException("Unable to close all file streams.", e);
-                return;
+                logException("Unable to close jsonGenerator and mapper file streams.", e);
             }
         }
+        super.close();
     }
 
     @RequiresApi(api = Build.VERSION_CODES.R)
@@ -237,15 +110,5 @@ public class JsonFileLogger implements FileLogger {
         } catch (IOException e) {
             logException("Unable to write antenna info to JSON", e);
         }
-    }
-
-    private void logException(String errorMessage, Exception e) {
-        Log.e(TAG, errorMessage, e);
-        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show();
-    }
-
-    private void logError(String errorMessage) {
-        Log.e(TAG, errorMessage);
-        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show();
     }
 }
